+++ packages/workspace-tools/src/executors/cargo-doc/schema.d.ts
@@ -1,0 +1,3 @@
+import { CargoBaseExecutorSchema } from "../../../declarations";
+
+export type CargoDocExecutorSchema = CargoBaseExecutorSchema;