+++ packages/workspace-tools/CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.159.1 (2024-08-26)
 
-
 ### Bug Fixes
 
-- **workspace-tools:** Resolved issues with the lint configuration files ([9f7d724c](https://github.com/storm-software/storm-ops/commit/9f7d724c))
+- **workspace-tools:** Resolved issues with the lint configuration files
+  ([9f7d724c](https://github.com/storm-software/storm-ops/commit/9f7d724c))
 
 ## 1.159.0 (2024-08-26)